Refactor admin endpoints into separate file
